# .eh_frame: Reuse unwind context and reduce copies