@javierhonduco 2022-12-14
> Request: If we really want to put those panics there, at least let's put a recover on the top level, return an error, and handle it gracefully. That way, we can have a compromise on this topic.

We currently use panics in `table.go` in places where a hard assertion makes sense, for example, when we know that some condition is impossible and should never happen.

Tried adding error handling which should probably be done at the iterator level, as it's where we continue parsing DWARF instructions and we may panic:

```diff
-func (ici *InstructionContextIterator) Next() *InstructionContext {
+func (ici *InstructionContextIterator) Next() (*InstructionContext, error) {
+       var theError error
+       defer func() {
+               if err := recover(); err != nil {
+                       theError = fmt.Errorf("panic occurred: %v", err)
+               }
+       }()
        for ici.ctx.buf.Len() > 0 {
```

Unfortunately the added function calls result in a significant regression:
```
[javierhonduco@fedora parca-agent]$ benchstat  /tmp/new  /tmp/new-catch-panics
name                                 old time/op    new time/op    delta
ParsingLibcUnwindInformation-12        3.34ms ±17%    4.06ms ± 9%  +21.59%  (p=0.000 n=29+30)
ParsingRedpandaUnwindInformation-12    37.7ms ± 2%    57.4ms ±11%  +52.48%  (p=0.000 n=30+30)

name                                 old alloc/op   new alloc/op   delta
ParsingLibcUnwindInformation-12        1.09MB ± 0%    1.09MB ± 0%   +0.00%  (p=0.000 n=28+29)
ParsingRedpandaUnwindInformation-12    14.1MB ± 0%    14.1MB ± 0%   +0.00%  (p=0.000 n=29+30)

name                                 old allocs/op  new allocs/op  delta
ParsingLibcUnwindInformation-12         41.8k ± 0%     41.8k ± 0%     ~     (all equal)
ParsingRedpandaUnwindInformation-12      527k ± 0%      527k ± 0%     ~     (all equal)
```

A possibility would be moving the `recover` outside of the parser. We'll have to make sure that we add it in the right places before we call the iterator in a loop. By reducing the cost of `defer`+ `anonymous function to recover` + `recover()` we won't have a regression compared to the original changeset in this PR. What do you think?