@javierhonduco 2022-12-16
All seems good:

```
[javierhonduco@fedora parca-agent]$ benchstat /tmp/main-6ad1efd /tmp/recover
name                                 old time/op    new time/op    delta
ParsingLibcUnwindInformation-12        11.5ms ±14%     3.2ms ±22%  -72.33%  (p=0.000 n=26+30)
ParsingRedpandaUnwindInformation-12    66.1ms ±22%    37.2ms ± 1%  -43.73%  (p=0.000 n=30+26)

name                                 old alloc/op   new alloc/op   delta
ParsingLibcUnwindInformation-12        4.63MB ± 0%    1.09MB ± 0%  -76.38%  (p=0.000 n=29+30)
ParsingRedpandaUnwindInformation-12    53.7MB ± 0%    14.1MB ± 0%  -73.78%  (p=0.000 n=30+29)

name                                 old allocs/op  new allocs/op  delta
ParsingLibcUnwindInformation-12         63.9k ± 0%     41.8k ± 0%  -34.58%  (p=0.000 n=30+30)
ParsingRedpandaUnwindInformation-12      783k ± 0%      527k ± 0%  -32.62%  (p=0.000 n=30+30)
```